:recycle: Separate edit cells from display cells
