+++ client/src/Main/TicketResultsTable/view.js
@@ -8,86 +8,13 @@ import {
   TableEditRow,
   TableEditColumn,
 } from '@devexpress/dx-react-grid-material-ui';
-import TableCell from '@material-ui/core/TableCell';
-import TextField from '@material-ui/core/TextField';
-import ProblemCategoryAutocomplete from '../Monitor/ProblemCategoryAutocomplete/ProblemCategoryAutocomplete';
-import { formatMilliSecondToTime, parseMillisecondFromFormattedTime } from 'Utils/TimeUtils';
 import { setFavicon } from 'Utils/FaviconUtils';
+import { Cell } from './Cell';
+import { EditCell } from './EditCell';
 
 import './style.css';
 
-const getRowId = row => row.id;
-
-const ProblemCategoryEditCell = ({ value, onValueChange }) => (
-  <TableCell>
-    <ProblemCategoryAutocomplete value={value || null} onChange={onValueChange} placeholder={''} />
-  </TableCell>
-);
-
-const DurationEditCell = ({ value, onValueChange }) => (
-  <TableCell>
-    <TextField
-      id="time"
-      type="time"
-      value={formatMilliSecondToTime(value)}
-      onChange={event => onValueChange(parseMillisecondFromFormattedTime(event.target.value))}
-      inputProps={{
-        step: 1,
-      }}
-    />
-  </TableCell>
-);
-
-const EditCell = props => {
-  const { column } = props;
-  if (column.name === 'problemCategory') {
-    return <ProblemCategoryEditCell {...props} />;
-  } else if (column.name === 'estimatedTime' || column.name === 'realTime') {
-    return <DurationEditCell {...props} />;
-  }
-  return <TableEditRow.Cell {...props} />;
-};
-
-const ProblemCategoryCell = ({ value, style }) => (
-  <TableCell style={style}>{value && value.description}</TableCell>
-);
-
-const DurationCell = ({ value, style, row }) => {
-  if (row.estimatedTime) {
-    style = {
-      ...style,
-      color: row.estimatedTime < row.realTime ? 'red' : 'green',
-    };
-  }
-
-  return <TableCell style={style}>{value && formatMilliSecondToTime(value)}</TableCell>;
-};
-
-const Cell = props => {
-  const { column, row } = props;
-  const style = row.addedOnTheFly
-    ? {
-        backgroundColor: '#ffe6e6',
-      }
-    : {};
-  if (column.name === 'problemCategory') {
-    return <ProblemCategoryCell {...props} style={style} />;
-  } else if (column.name === 'estimatedTime' || column.name === 'realTime') {
-    return <DurationCell {...props} style={style} />;
-  }
-  return <Table.Cell {...props} style={style} />;
-};
-
 class TicketResultsTable extends React.Component {
-  componentDidMount() {
-    document.title = this.title;
-    setFavicon('caspr');
-  }
-
-  componentWillUnmount() {
-    document.title = 'Caspr';
-  }
-
   title = '#' + this.props.ticketData.thirdPartyId + ' ' + this.props.ticketData.description;
 
   state = {
@@ -99,6 +26,16 @@ class TicketResultsTable extends React.Component {
       { name: 'problemCategory', title: 'Problem Category' },
     ],
   };
+
+  componentDidMount() {
+    document.title = this.title;
+    setFavicon('caspr');
+  }
+
+  componentWillUnmount() {
+    document.title = 'Caspr';
+  }
+
   commitChanges = ({ changed }) => {
     if (changed) {
       const tasksToUpdate = this.props.ticketData.tasks
@@ -122,7 +59,7 @@ class TicketResultsTable extends React.Component {
             <Grid
               rows={this.props.ticketData.tasks}
               columns={this.state.columns}
-              getRowId={getRowId}
+              getRowId={row => row.id}
             >
               <EditingState onCommitChanges={this.commitChanges} />
               <Table cellComponent={Cell} />
